Add color converters, schema/index updates, and lipsum demo
- Added WPF value converters for trade status/direction and booleans, with global registration in Converters.xaml and App.xaml.
- Enhanced MainWindow.xaml UI with color cues and improved header layout.
- Changed Trade.sql to store P&L and fee fields as INTEGER, and updated indexes to use composite columns for better query performance.
- Updated SqliteScriptsGen.settings to generate new composite indexes.
- Bumped all project versions to 0.1.7.
- Improved demo data seeding: appended demo notes and generated review notes using new Text.cs and Text.lipsum.cs helpers.
- Fixed sample trade direction in MockTradeHistoryVM.
- Minor XAML margin adjustments for notes fields.
- Adjusted codegen order in OzzTradeDiary.OzzGen for correct output.

+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/DesignTime/MockTradeHistoryVM.cs
@@ -17,7 +17,7 @@ namespace TD.WPF.DesignTime
                 SymbolId = 1,
                 MarketType = MarketType.CryptoPerpetual,
                 EntryMethod = EntryMethod.Limit,
-                TradeDirection = TradeDirection.Long,
+                TradeDirection = TradeDirection.Short,
                 TradeStatus = TradeStatus.Closed,
                 Tags = "breakout,high-rr",
                 EntryTime = DateTime.Now.AddDays(-10),
